Add images and move components into proper component folder
