+++ src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import Header from "./Header";
+import Header from "./components/Header";
 import Home from "./pages/Home";
 import About from "./pages/About";
